strto_xxx refactoring
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@527 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ partition.h
@@ -46,6 +46,6 @@ int consistencyCheck(struct partition *partTable, int doprint, int verbose,
 		     struct device *used_dev, int target_partition);
 
 void setBeginEnd(struct partition *partTable,
-		 unsigned int begin, unsigned int end,
+		 unsigned long begin, unsigned long end,
 		 unsigned int heads, unsigned int sector,
 		 int activate, int type, int fat_bits);
